+++ src/FolkerKinzel.VCards/Models/DataProperty.cs
@@ -200,20 +200,23 @@ public abstract class DataProperty : VCardProperty, IEnumerable<DataProperty>
 
     internal static DataProperty Parse(VcfRow vcfRow, VCdVersion version)
     {
-        if (DataUrl.TryParse(vcfRow.Value, out DataUrlInfo info))
+        if (DataUrl.TryParse(vcfRow.Value, out DataUrlInfo dataUrlInfo))
         {
+            ReadOnlyMemory<char> mime = dataUrlInfo.MimeType;
+            UnMaskMimeType(ref mime);
+
             vcfRow.Parameters.MediaType =
-                MimeTypeInfo.TryParse(info.MimeType, out MimeTypeInfo mimeTypeInfo)
+                MimeTypeInfo.TryParse(mime, out MimeTypeInfo mimeTypeInfo)
                                   ? mimeTypeInfo.ToString()
                                   : MimeString.OctetStream;
 
-            return info.TryGetEmbeddedData(out OneOf<string, byte[]> data)
+            return dataUrlInfo.TryGetEmbeddedData(out OneOf<string, byte[]> data)
                     ? data.Match<DataProperty>
                        (
                         s => new EmbeddedTextProperty(new TextProperty(vcfRow, version)),
                         b => new EmbeddedBytesProperty(b, vcfRow.Group, vcfRow.Parameters)
                         )
-                    : FromText(vcfRow.Value.ToString(), info.MimeType.ToString(), vcfRow.Group);
+                    : FromText(vcfRow.Value.ToString(), dataUrlInfo.MimeType.ToString(), vcfRow.Group);
         }
 
         if (vcfRow.Parameters.Encoding == Enc.Base64)
@@ -267,6 +270,27 @@ public abstract class DataProperty : VCardProperty, IEnumerable<DataProperty>
         }
     }
 
+    private static void UnMaskMimeType(ref ReadOnlyMemory<char> mime)
+    {
+        int trimEndLength = 0;
+        ReadOnlySpan<char> span = mime.Span;
+
+        if(span.EndsWith(@";base64\")) // masked comma
+        {
+            trimEndLength += 8;
+            span = span.Slice(0, span.Length - trimEndLength);
+
+            if(span.EndsWith('\\')) // masked semicolon
+            {
+                trimEndLength++;
+                span = span.Slice(0, span.Length - 1);
+            }
+
+            // The MIME type may have parameters, separated by masked semicolons:
+            mime = span.Contains('\\') ? span.UnMaskValue(VCdVersion.V4_0).AsMemory() 
+                                       : mime.Slice(0, mime.Length - trimEndLength);
+        }
+    }
 
     private void InitializeValue()
     {