+++ src/FolkerKinzel.VCards/md/ReleaseNotes.md
@@ -1,5 +1,6 @@
 - Supports RFC 6868 parameter value encoding in vCard 4.0
 - Fixes an issue that the properties `FBURL`, `CALURI`, `CAPURI`, and `CALADRURI` in vCard 3.0 might have a 
 PREF parameter whose TYPE parameter name is missing. (This was caused by an [error in RFC 2739](https://www.rfc-editor.org/errata/eid7914)).
+- Fixes an issue that the `MediaType` of a "data" URL could not be parsed correctly in vCard 4.0 if masking had been applied to the "data" URL.
 &nbsp;
 >**Project reference:** On some systems the content of the CHM file in the Assets is blocked. Before opening the file right click on the file icon, select Properties, and **check the "Allow" checkbox** - if it is present - in the lower right corner of the General tab in the Properties dialog.